Fix bugs caused by renaming
